+++ src/api/git.js
@@ -1,11 +1,16 @@
-const { error } = require('console');
 const { spawn } = require('child_process');
+const { error } = require('console');
 const { makeArch } = require('./zip');
 
-module.exports = {
-  cloneSkeleton: async (address, folderName) => {
-    const process = await spawn('git', ['clone', address]);
-    process.on('close', () => makeArch('skeleton', folderName));
-    process.on('error', err => error(err));
-  },
+/**
+ * Start process cloning repository from github
+ * @param {String} address
+ * @param {String} folderName
+ */
+const cloneSkeleton = (address, folderName) => {
+  const process = spawn('git', ['clone', address]);
+  process.on('close', () => makeArch('skeleton', folderName));
+  process.on('error', err => error(err));
 };
+
+module.exports = cloneSkeleton;