+++ src/api/zip.js
@@ -1,10 +1,14 @@
 /* eslint-disable class-methods-use-this */
 const Unzip = require('adm-zip');
+const { log } = require('console');
 const { resolve } = require('path');
-const { error, log } = require('console');
 const { mkdir, remove } = require('fs-extra');
 
 class Zip {
+  /**
+   * Exact arch in folder
+   * @param {String} pathToArch
+   */
   async exactArch(pathToArch) {
     const zip = new Unzip(resolve(__dirname, pathToArch));
 
@@ -12,6 +16,11 @@ class Zip {
     await zip.extractAllTo('./qa-skeleton', true);
   }
 
+  /**
+   * Create archive with test env
+   * @param {String} name
+   * @param {String} folderName
+   */
   async makeArch(name, folderName) {
     const zip = new Unzip();
 
@@ -19,8 +28,8 @@ class Zip {
     await zip.writeZip(`${name}.zip`);
 
     await remove(folderName, (err) => {
-      if (err) error(err);
-      log('EXCELLENT');
+      if (err) throw new Error(err);
+      log('Success');
     });
   }
 }