+++ src/cli/files.js
@@ -1,22 +1,19 @@
 const { prompt } = require('inquirer');
-const { resolve } = require('path');
-const { copyFile } = require('fs-extra');
-const { error, log } = require('console');
+const { emptyTemplateDoc, getTemplate, move } = require('../api/files');
 
 const question = [
   {
     type: 'list',
     name: 'docker',
     message: 'Select file to generate',
-    choices: ['docker-simple', 'docker-with-noVNC', 'docker-three-node-noVNC', 'chrome-config'],
+    choices: [
+      'docker-simple', 'docker-with-noVNC',
+      'chrome-config',
+      'doc',
+      'xlsx',
+    ],
   },
 ];
-
-const move = (fileName, afterName) => copyFile(resolve(__dirname, fileName), afterName,
-  (err) => {
-    if (err) error(err);
-    log('Success');
-  });
 
 module.exports = () => {
   prompt(question)
@@ -28,14 +25,16 @@ module.exports = () => {
         case 'docker-with-noVNC':
           move('docker-compose-novnc.yml', 'qa-docker-compose.yml');
           break;
-        case 'docker-three-node-noVNC':
-          move('docker-compose.yml', 'qa-docker-compose.yml');
-          break;
         case 'chrome-config':
           move('../assets/config/index.js', 'chrome.config.js');
           break;
+        case 'doc':
+          emptyTemplateDoc('../assets/docs/test-case-template.doc', 'doc');
+          break;
+        case 'xlsx':
+          getTemplate();
+          break;
         default:
-          log('Pleas try again');
           break;
       }
     });