+++ src/cli/generator.js
@@ -1,4 +1,3 @@
-const { log } = require('console');
 const { prompt } = require('inquirer');
 const { exactArch } = require('../api/zip');
 
@@ -7,13 +6,7 @@ const questions = [
     type: 'list',
     name: 'type',
     message: 'Select the test skeleton to generate:',
-    choices: [
-      'selenium',
-      'appium',
-      'wdio',
-      'selenium+jest',
-      'locus',
-    ],
+    choices: ['selenium', 'appium', 'selenium+jest', 'locus', 'python+selenium'],
   },
 ];
 
@@ -27,19 +20,17 @@ module.exports = () => {
         case 'appium':
           exactArch('../assets/appiumwdblui.zip');
           break;
-        case 'wdio':
-          exactArch('../assets/selwdioblui.zip');
-          break;
         case 'selenium+jest':
           exactArch('../assets/selje.zip');
           break;
         case 'locus':
           exactArch('../assets/locus.zip');
           break;
+        case 'python+selenium':
+          exactArch('../assets/pyT.zip');
+          break;
         default:
-          log('Something error');
           break;
       }
-    })
-    .catch();
+    }).catch();
 };