+++ src/cli/index.js
@@ -1,25 +1,12 @@
 const commander = require('commander');
-const { log } = require('console');
-
 const gitCli = require('./git');
-const dockerCli = require('./files');
-const documentCli = require('./document');
+const filesCli = require('./files');
 const generatorCli = require('./generator');
 
 commander
-  .command('v')
-  .description('version')
-  .action(() => log('v 2.0'));
-
-commander
   .command('file:gen')
-  .description('Command generate docker files')
-  .action(() => dockerCli());
-
-commander
-  .command('doc:gen')
-  .description('Command generate documents')
-  .action(() => documentCli());
+  .description('Command generate files')
+  .action(() => filesCli());
 
 commander
   .command('git:cli')
@@ -31,4 +18,7 @@ commander
   .description('Call cli for generate skeleton')
   .action(() => generatorCli());
 
-module.exports = commander;
+module.exports = () => {
+  commander.parse(process.argv);
+  if (commander.args.length === 0) commander.help();
+};