+++ test/git.spec.js
@@ -1,4 +1,4 @@
-const clone = require('download-git-repo');
+const cloneSkeleton = require('../lib/api/git');
 
 const { error } = require('console');
 const { assert } = require('chai');
@@ -8,7 +8,7 @@ describe('Test git api', () => {
   before(() => mkdir('./test/tmp'));
 
   it('test clone repository', async () => {
-    await clone('AndriyNikiforov/appiumwdblui', './test/tmp', (err) => {
+    await cloneSkeleton('https://github.com/AndriyNikiforov/appiumwdblui.git', 'appiumwdblui').then((err) => {
       if (err) error(err);
       access('./test/tmp/README.md', error => assert.isNull(error));
       remove('./test/tmp');